Fixed PyPi issues. Ready for code review.
