--- python-port/sendpy/sendpy.egg-info/PKG-INFO
@@ -1,30 +1,0 @@
-Metadata-Version: 2.1
-Name: sendpy
-Version: 0.1
-Summary: Email and text notification program
-Home-page: https://github.com/tdulcet/Send-Msg-CLI
-Author: Daniel Connelly
-Author-email: connellyd2050@gmail.com
-License: GNU
-Description: sendpy
-        --------
-        # Libraries Used (Suggestion credit: Teal Dulcet)
-        [email](https://docs.python.org/3/library/email.html)
-        [smptlib](https://docs.python.org/3/library/smtplib.html)
-        
-        ## Ways to Email Oneself
-        
-        # With script
-        
-        output=$(sudo python3 program.py 2>&1); sendpy.py "'$HOSTNAME' is done"'!' "Exit code: $?\nOutput:\n$output\n"
-        [comment]: # ( output=$(sudo python2 runMythNew.py 2>&1); ./send.sh "'$HOSTNAME' is done"'!' "Exit code: $?\nOutput:\n$output\n" )
-        
-        
-Keywords: sendpy email e-mail send text notification g-mail gmail
-Platform: UNKNOWN
-Classifier: Programming Language :: Python :: 3
-Classifier: License :: OSI Approved :: MIT License
-Classifier: Operating System :: OS Independent
-Classifier: Development Status :: 1 - Planning
-Requires-Python: >=3.0
-Description-Content-Type: text/markdown
