--- python-port/sendpy/sendpy.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
