--- python-port/sendpy/sendpy.egg-info/not-zip-safe
@@ -1,1 +1,0 @@
-
