--- python-port/sendpy/sendpy_danc2.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-sendpy
