--- python-port/sendpy/sendpy_danc2050.egg-info/SOURCES.txt
@@ -1,14 +1,0 @@
-MANIFEST.in
-README.rst
-setup.py
-bin/sendpy
-sendpy/__init__.py
-sendpy/__main__.py
-sendpy/configuration.py
-sendpy/send.py
-sendpy/usage.py
-sendpy_danc2050.egg-info/PKG-INFO
-sendpy_danc2050.egg-info/SOURCES.txt
-sendpy_danc2050.egg-info/dependency_links.txt
-sendpy_danc2050.egg-info/not-zip-safe
-sendpy_danc2050.egg-info/top_level.txt
