--- python-port/sendpy/sendpy_danc2050.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
