--- python-port/sendpy/sendpy_danc2050.egg-info/not-zip-safe
@@ -1,1 +1,0 @@
-
